feat: Add function to reverse a linked list
This commit adds a new function `reverseList` to the code. The function takes the head of a linked list as input and reverses the order of the nodes in the list. It uses a iterative approach to reverse the links between the nodes.

The commit also includes an example usage of the `reverseList` function, creating a linked list with values [1, 2, 3, 4, 5] and then reversing it.

The purpose of this change is to provide a utility function that can be used to reverse linked lists in various scenarios.

```